--- DBMS/CTableEntity.cpp
@@ -1,4 +1,0 @@
-#include "stdafx.h"
-#include "CTableEntity.h"
-
-CTableEntity::~CTableEntity(){}
